Update inf_attn for flash_lobo exploration
This sets up an exploration for infinite attention (no limit on # heads)

Before, we experimented with this and the value-head dimension, as
higher values would lead to better capacity.

The potential benefit of inf_cproj is that we have a lot of space in
the residual, however, we limit our cproj number by the residual
dimesion.

As we actually have more dimensions than the residual dimension, we can
safely pack it, due to JL.

However, if attention heads are adding noise, we might be able to see a
greater benefit of more heads if we allow heads to suppress this noise
wiht the learned-off-by-one offsets.

To check, we just need to add flash lobo, and see if either or both
appear to unlimit benefits from adding more attention heads.

Note: Another test, would be to check if we can use infinite queries,
with just one kv group with flash_lobo and higher v_head_dimension, but
will leave this to a follow-up commit,