Add init scale factor
Seems that the alternative scaling factors only work if on the order of
magnitude as the prior gaussian weights. This may be because the
remaining weights are also at this scale.

Creating test to determine if the scale needs to match, and what ratio
the scaling factor needs to be for each initialization.