MC datasets ensure coverage of all possible tokens
This updates the char_convert.py which is now able to emit a token list,
which contains every possible output token.

This then is used to create a meta.pkl file with the prepare.py script
with all of the possible chars -- not just those within the dataset.

Finally run the last prepare.py tokenization with --reuse_chars option
in order to ensure we have all of the possible tokens allowed.